requirements.txt
altgraph==0.17.4
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.24
debugpy==1.8.0
decorator==5.1.1
easyocr==1.7.0
entrypoints==0.4
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
fsspec==2023.9.2
gast==0.3.3
ghostscript==0.7
idna==3.4
imageio==2.31.1
imagesize==1.4.1
imgaug==0.4.0
ipykernel==6.16.2
ipyparallel==8.6.1
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
numpy==1.26.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==1.3.5
parso==0.8.3
pdf2image==1.16.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
pyclipper==1.3.0.post5
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
queues==0.6.3
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.6
Wand==0.6.11
wcwidth==0.2.8
Werkzeug==3.0.0
